Added explosion tests
We now have continuous single explosion tests, and one-time 100 explosion tests.

Improved bounds checks to immediately remove any object that is entirely within the explosion radius.

Improved SubtractBulk slightly by using a dictionary of input group points to their input group indices.

Added TrueBulkExplosionExecutor which sends all polygons and explosions to the subtractor at once, but doesn't do any bounds checking.

Performance is about the same for the iterative and bulk approaches, but true bulk is extremely slow, except in the case that all objects are affected by explosions anyway, which produces similar performance to the others.

For this reason, iterative is still the best option.