O'Rourke works with basic convex unit tests (degenerate cases and holes not yet tested)
Reversed all polygon orientations! They are now CCW as expected, with CW holes. It turns out that I had Clipper set up to take the clipping polygons in the negative orientation, not positive, but the output was positive anyway.

Polygons can now be compared for equality! This is crucial for unit tests and gives me a lot more power.